+++ atcg_lib/src/Network/TCPClient.cpp
@@ -53,11 +53,11 @@ void TCPClient::disconnect()
     impl->connected = false;
 }
 
-torch::Tensor TCPClient::sendAndWait(uint8_t* data)
+std::vector<uint8_t> TCPClient::sendAndWait(const std::vector<uint8_t>& data)
 {
-    uint32_t read_offset = 0;
-    uint32_t data_size   = atcg::NetworkUtils::readInt<uint32_t>(data, read_offset);
-    if(impl->socket.send(data, data_size + sizeof(uint32_t)) != sf::Socket::Status::Done)
+    sf::Packet packet;
+    packet.append(data.data(), data.size());
+    if(impl->socket.send(packet) != sf::Socket::Status::Done)
     {
         ATCG_ERROR("Could not send data...");
         return {};
@@ -76,17 +76,14 @@ torch::Tensor TCPClient::sendAndWait(uint8_t* data)
     } while(total_received < sizeof(uint32_t));
 
     uint32_t expected_size = atcg::hton<uint32_t>(message_size);
-    torch::Tensor rec_data = torch::empty({(int)expected_size}, atcg::TensorOptions::uint8HostOptions());
-    total_received         = 0;
+    std::vector<uint8_t> rec_data(expected_size);
+    total_received = 0;
     while(total_received < expected_size)
     {
-        impl->socket.receive((char*)((uint8_t*)rec_data.data_ptr() + total_received),
-                             rec_data.numel() - total_received,
-                             received);
+        impl->socket.receive((char*)(rec_data.data() + total_received), expected_size - total_received, received);
         total_received += received;
     }
 
-    // Zero terminate data
     return rec_data;
 }
 }    // namespace atcg