+++ tests/Network_test.cpp
@@ -491,22 +491,21 @@ TEST(NetworkTest, forceDisconnect)
 
 TEST(NetworkTest, simpleEcho)
 {
-    uint8_t* send_buffer = new uint8_t[1024];
+    std::vector<uint8_t> send_buffer(4);
 
     uint32_t write_offset = 0;
-    atcg::NetworkUtils::writeInt<uint32_t>(send_buffer, write_offset, sizeof(uint32_t));
-    atcg::NetworkUtils::writeInt<uint32_t>(send_buffer, write_offset, 42);
+    atcg::NetworkUtils::writeInt<uint32_t>(send_buffer.data(), write_offset, 42);
 
     atcg::TCPServer server;
     atcg::TCPClient client;
     server.setOnReceiveCallback(
-        [&](uint8_t* data, uint32_t data_size, uint64_t client_id)
+        [&](std::vector<uint8_t>& data, uint64_t client_id)
         {
             uint32_t read_offset = 0;
-            uint32_t message     = atcg::NetworkUtils::readInt<uint32_t>(data, read_offset);
-            EXPECT_EQ(data_size, sizeof(uint32_t));
+            uint32_t message     = atcg::NetworkUtils::readInt<uint32_t>(data.data(), read_offset);
+            EXPECT_EQ(data.size(), sizeof(uint32_t));
             EXPECT_EQ(message, 42);
-            server.sendToClient(data, data_size, client_id);
+            server.sendToClient(data, client_id);
         });
     server.start("127.0.0.1", 25565);
 
@@ -514,7 +513,7 @@ TEST(NetworkTest, simpleEcho)
     auto response = client.sendAndWait(send_buffer);
 
     uint32_t read_offset = 0;
-    EXPECT_EQ(atcg::NetworkUtils::readInt<uint32_t>(response.data_ptr<uint8_t>(), read_offset), 42);
+    EXPECT_EQ(atcg::NetworkUtils::readInt<uint32_t>(response.data(), read_offset), 42);
 
     client.disconnect();
     server.stop();
@@ -522,23 +521,22 @@ TEST(NetworkTest, simpleEcho)
 
 TEST(NetworkTest, streamEcho)
 {
-    uint8_t* send_buffer = new uint8_t[1024];
+    std::vector<uint8_t> send_buffer(1024);
     atcg::TCPServer server;
     atcg::TCPClient client;
-    server.setOnReceiveCallback([&](uint8_t* data, uint32_t data_size, uint64_t client_id)
-                                { server.sendToClient(data, data_size, client_id); });
+    server.setOnReceiveCallback([&](std::vector<uint8_t>& data, uint64_t client_id)
+                                { server.sendToClient(data, client_id); });
     server.start("127.0.0.1", 25565);
 
     client.connect("127.0.0.1", 25565);
     for(uint32_t i = 0; i < 1024; ++i)
     {
         uint32_t write_offset = 0;
-        atcg::NetworkUtils::writeInt<uint32_t>(send_buffer, write_offset, sizeof(uint32_t));
-        atcg::NetworkUtils::writeInt<uint32_t>(send_buffer, write_offset, i);
+        atcg::NetworkUtils::writeInt<uint32_t>(send_buffer.data(), write_offset, i);
 
         auto response        = client.sendAndWait(send_buffer);
         uint32_t read_offset = 0;
-        EXPECT_EQ(atcg::NetworkUtils::readInt<uint32_t>(response.data_ptr<uint8_t>(), read_offset), i);
+        EXPECT_EQ(atcg::NetworkUtils::readInt<uint32_t>(response.data(), read_offset), i);
     }
 
     client.disconnect();
@@ -547,30 +545,24 @@ TEST(NetworkTest, streamEcho)
 
 TEST(NetworkTest, echoLargeData)
 {
-    int data_size        = 10000;
-    auto buffer          = torch::randn({data_size}, atcg::TensorOptions::floatHostOptions());
-    uint8_t* send_buffer = new uint8_t[buffer.numel() * buffer.element_size() + sizeof(uint32_t)];
-
-    uint32_t write_offset = 0;
-    atcg::NetworkUtils::writeBuffer(send_buffer,
-                                    write_offset,
-                                    (uint8_t*)buffer.data_ptr(),
-                                    buffer.numel() * buffer.element_size());
+    int data_size = 10000;
+    auto buffer   = torch::randn({data_size}, atcg::TensorOptions::floatHostOptions());
 
     atcg::TCPServer server;
     atcg::TCPClient client;
-    server.setOnReceiveCallback([&](uint8_t* data, uint32_t data_size, uint64_t client_id)
-                                { server.sendToClient(data, data_size, client_id); });
+    server.setOnReceiveCallback([&](std::vector<uint8_t>& data, uint64_t client_id)
+                                { server.sendToClient(data, client_id); });
     server.start("127.0.0.1", 25565);
 
     client.connect("127.0.0.1", 25565);
-    auto response = client.sendAndWait(send_buffer);
+    auto response = client.sendAndWait(
+        std::vector<uint8_t>((uint8_t*)buffer.data_ptr(), (uint8_t*)(buffer.data_ptr<float>() + buffer.numel())));
 
-    float* float_ptr = reinterpret_cast<float*>(response.data_ptr());
+    float* float_ptr = reinterpret_cast<float*>(response.data());
 
-    response = atcg::createHostTensorFromPointer(float_ptr, {data_size}).clone();
+    auto response_tnsr = atcg::createHostTensorFromPointer(float_ptr, {data_size}).clone();
 
-    EXPECT_EQ(torch::sum(response - buffer).item<float>(), 0);
+    EXPECT_EQ(torch::sum(response_tnsr - buffer).item<float>(), 0);
 
     client.disconnect();
     server.stop();
@@ -578,30 +570,24 @@ TEST(NetworkTest, echoLargeData)
 
 TEST(NetworkTest, echoExtraLargeData)
 {
-    int data_size        = 5000000;
-    auto buffer          = torch::randn({data_size}, atcg::TensorOptions::floatHostOptions());
-    uint8_t* send_buffer = new uint8_t[buffer.numel() * buffer.element_size() + sizeof(uint32_t)];
-
-    uint32_t write_offset = 0;
-    atcg::NetworkUtils::writeBuffer(send_buffer,
-                                    write_offset,
-                                    (uint8_t*)buffer.data_ptr(),
-                                    buffer.numel() * buffer.element_size());
+    int data_size = 5000000;
+    auto buffer   = torch::randn({data_size}, atcg::TensorOptions::floatHostOptions());
 
     atcg::TCPServer server;
     atcg::TCPClient client;
-    server.setOnReceiveCallback([&](uint8_t* data, uint32_t data_size, uint64_t client_id)
-                                { server.sendToClient(data, data_size, client_id); });
+    server.setOnReceiveCallback([&](std::vector<uint8_t>& data, uint64_t client_id)
+                                { server.sendToClient(data, client_id); });
     server.start("127.0.0.1", 25565);
 
     client.connect("127.0.0.1", 25565);
-    auto response = client.sendAndWait(send_buffer);
+    auto response = client.sendAndWait(
+        std::vector<uint8_t>((uint8_t*)buffer.data_ptr(), (uint8_t*)(buffer.data_ptr<float>() + buffer.numel())));
 
-    float* float_ptr = reinterpret_cast<float*>(response.data_ptr());
+    float* float_ptr = reinterpret_cast<float*>(response.data());
 
-    response = atcg::createHostTensorFromPointer(float_ptr, {data_size}).clone();
+    auto response_tnsr = atcg::createHostTensorFromPointer(float_ptr, {data_size}).clone();
 
-    EXPECT_EQ(torch::sum(response - buffer).item<float>(), 0);
+    EXPECT_EQ(torch::sum(response_tnsr - buffer).item<float>(), 0);
 
     client.disconnect();
     server.stop();
@@ -609,41 +595,35 @@ TEST(NetworkTest, echoExtraLargeData)
 
 TEST(NetworkTest, calculateExtraLargeData)
 {
-    int data_size        = 5000000;
-    auto buffer          = torch::randn({data_size}, atcg::TensorOptions::floatHostOptions());
-    uint8_t* send_buffer = new uint8_t[buffer.numel() * buffer.element_size() + sizeof(uint32_t)];
-
-    uint32_t write_offset = 0;
-    atcg::NetworkUtils::writeBuffer(send_buffer,
-                                    write_offset,
-                                    (uint8_t*)buffer.data_ptr(),
-                                    buffer.numel() * buffer.element_size());
+    int data_size = 5000000;
+    auto buffer   = torch::randn({data_size}, atcg::TensorOptions::floatHostOptions());
 
     atcg::TCPServer server;
     atcg::TCPClient client;
     server.setOnReceiveCallback(
-        [&](uint8_t* data, uint32_t data_size, uint64_t client_id)
+        [&](std::vector<uint8_t>& data, uint64_t client_id)
         {
-            float* float_ptr    = (float*)data;
-            uint32_t num_floats = data_size / sizeof(float);
+            float* float_ptr    = (float*)data.data();
+            uint32_t num_floats = data_size;
 
             for(int i = 0; i < num_floats; ++i)
             {
                 float_ptr[i] *= 2.0f;
             }
 
-            server.sendToClient(data, data_size, client_id);
+            server.sendToClient(data, client_id);
         });
     server.start("127.0.0.1", 25565);
 
     client.connect("127.0.0.1", 25565);
-    auto response = client.sendAndWait(send_buffer);
+    auto response = client.sendAndWait(
+        std::vector<uint8_t>((uint8_t*)buffer.data_ptr(), (uint8_t*)(buffer.data_ptr<float>() + buffer.numel())));
 
-    float* float_ptr = reinterpret_cast<float*>(response.data_ptr());
+    float* float_ptr = reinterpret_cast<float*>(response.data());
 
-    response = atcg::createHostTensorFromPointer(float_ptr, {data_size}).clone();
+    auto response_tnsr = atcg::createHostTensorFromPointer(float_ptr, {data_size}).clone();
 
-    EXPECT_EQ(torch::sum(response - 2.0f * buffer).item<float>(), 0);
+    EXPECT_EQ(torch::sum(response_tnsr - 2.0f * buffer).item<float>(), 0);
 
     client.disconnect();
     server.stop();